+++ app/Http/Controllers/API/AuthController.php
@@ -141,10 +141,9 @@ class AuthController extends Controller
 
 
 
-    public function logout()
+    public function logout(Request $request)
     {
         if (Auth::check()) {
-            //Auth::logout();
             Auth::user()->tokens()->delete();
 
             return response()->json([
@@ -182,179 +181,5 @@ class AuthController extends Controller
         }
 
         return $randomString;
-    }
-
-
-    /**
-     * Validate the request data and send a verification code to the provided email address.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\JsonResponse
-     */
-    public function email_send_code(Request $request)
-    {
-        // Validate the request data
-        $request->validate([
-            'email' => 'required|string|email|max:255',
-            'password' => 'required|string'
-        ]);
-
-        // Get the authenticated user
-        $user = Auth::user();
-
-        // Check if the user is not authenticated or the password is incorrect
-        if (!$user || !Hash::check($request->input('password'), $user->password)) {
-            return response()->json([
-                'message' => 'Unauthenticated',
-            ], Response::HTTP_UNAUTHORIZED);
-        }
-
-        // Check if the email is already registered for another user
-        $existingUser = User::where('email', $request->email)->where('id', '!=', $user->id)->first();
-
-        if ($existingUser) {
-            return response()->json([
-                'message' => 'Email already registered',
-            ], Response::HTTP_CONFLICT);
-        }
-
-        // Create or update the email verification code for the user
-        $emailToVerify = $user->emailToVerify ?? new EmailToVerify;
-        $emailToVerify->email = $request->email;
-        $emailToVerify->code = self::str_random(10);
-        $user->emailToVerify()->save($emailToVerify);
-
-        // Send the verification code to the provided email address
-        SendMailCodeJob::dispatch($request->email, $user->name, $emailToVerify->code);
-        //Mail::to($request->email)->send(new ConfirmationCode($emailToVerify->code, $user->name));
-
-        // Return a success response
-        return response()->json([
-            'message' => 'Email sent',
-        ], Response::HTTP_OK);
-    }
-    // verifica el email con el codigo obtenido.
-    public function verify_code_email(Request $request)
-    {
-        $request->validate(['code' => 'required|regex:/^[A-Z0-9]{10}$/', 'email' => 'required|string|email|max:255']);
-        $code = $request->code;
-        if (Auth::check()) {
-            $user = Auth::user();
-            $emailToVerify = $user->emailToVerify;
-
-            if (!$emailToVerify || $emailToVerify->code !== $request->code) {
-                return response()->json(['error' => 'Invalid code'], Response::HTTP_BAD_REQUEST);
-            }
-
-            $expirationTime = now()->subMinutes(5);
-            if ($emailToVerify->updated_at < $expirationTime) {
-                return response()->json(['error' => 'Verification code has expired'], Response::HTTP_BAD_REQUEST);
-            }
-
-            if ($request->user()->markEmailAsVerified()) {
-                $user->email = $emailToVerify->email;
-                $emailToVerify->delete();
-                $user->save();
-                //event(new Verified($user));
-                return response()->json(['message' => 'Confirmation successful'], Response::HTTP_OK);
-            } else {
-                return response()->json(['error' => 'Failed to mark email as verified'], Response::HTTP_INTERNAL_SERVER_ERROR);
-            }
-
-        } else {
-            return response()->json(['error' => 'Unauthenticated'], Response::HTTP_UNAUTHORIZED);
-        }
-    }
-
-    // genera una peticion de cambio de contraseña , solo 1 cada 3 minutos
-    public function forgot_password(Request $request)
-    {
-        $request->validate([
-            'cuil' => [
-                'required',
-                'string',
-                'regex:/^\d{2}-\d{8}-\d{1}$/',
-                'exists:users,cuil'
-            ]
-        ]);
-
-        $user = User::where('cuil', $request->cuil)->first();
-
-        if (!$user) {
-            return response()->json(['message' => 'Invalid cuil provided.'], Response::HTTP_BAD_REQUEST);
-        }
-        $expirationTime = now()->subMinutes(5);
-        if ($user->forgotpassword && $user->forgotpassword->created_at > $expirationTime) {
-            return response()->json(['message' => 'Wait 5 minutes and try again'], Response::HTTP_BAD_REQUEST);
-        }
-
-        $TOKEN = Str::random(20);
-        $passwordReset = $user->forgotpassword()->updateOrCreate(
-            ['email' => $request->email],
-            ['token' => Hash::make($TOKEN)]
-        );
-
-        // Send email with the new password
-        SendMailForgotPassJob::dispatch($request->email, $user->name, $TOKEN);
-        //Mail::to($user->email)->send(new ForgotPassMail($user->name, $TOKEN));
-
-        return response()->json(['message' => 'Password reset successful. Check your email.'], Response::HTTP_OK);
-    }
-    // el commit de la peticion de cambio de contraseña , asigna la nueva y borra la paticion.
-    public function reset_password(Request $request)
-    {
-        $request->validate([
-            'cuil' => 'required|email|exists:users,email',
-            'token' => 'required|string',
-            'password' => 'required|string'
-        ]);
-
-        $user = User::where('cuil', $request->email)->first();
-
-        if (!$user) {
-            return response()->json(['message' => 'User not found.'], Response::HTTP_NOT_FOUND);
-        }
-
-        $passwordReset = $user->forgotpassword();
-
-        if (!$passwordReset) {
-            return response()->json(['message' => 'Password reset token not found.'], Response::HTTP_NOT_FOUND);
-        }
-
-        if (!Hash::check($request->token, $passwordReset->token)) {
-            return response()->json(['message' => 'Invalid token.'], Response::HTTP_UNAUTHORIZED);
-        }
-
-        // Update user's password
-        $user->password = Hash::make($$request->password);
-        $user->save();
-
-        // Delete the password reset token
-        $passwordReset->delete();
-
-        return response()->json(['message' => 'Password reset successful. Check your email for the new password.'], Response::HTTP_OK);
-    }
-    //Change password , si estas logado y conoces tu  ocntraseña anterior
-    public function change_password(Request $request)
-    {
-        $request->validate([
-            'current_password' => 'required',
-            'new_password' => 'required|min:8|different:current_password|confirmed|regex:/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^a-zA-Z\d]).+$/',
-        ]);
-        if (Auth::check()) {
-            $user = Auth::user();
-            if (!Hash::check($request->input('current_password'), $user->password)) {
-                return response()->json(['error' => 'Current password is incorrect'], Response::HTTP_BAD_REQUEST);
-            }
-            $user->forceFill([
-                'password' => Hash::make($request->input('new_password')),
-            ]);
-            $user->save();
-            //Auth::logout();
-            $user->tokens()->delete();
-            return response()->json(['message' => 'Password changed successfully'], Response::HTTP_OK);
-        }
-
-        return response()->json(['error' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
     }
 }
