+++ config/auth.php
@@ -114,6 +114,6 @@ return [
     |
     */
 
-    'password_timeout' => 10800,
+    'password_timeout' => 60,
 
 ];
